Moving 3 tests that fill out the entire form and expect PDFs to be generated, so now we only have one (testing both the confirmation page, custom receipt and regular receipt). This should save us some time. Also adding support for the 'createPdf' cookie that needs to be set to for the backend to generate PDFs while filling out the form. This is fairly slow, so we should minimize it.
